improve the config color option, set hex color value(RGB888 format) directly

+++ TFT/src/User/API/Colors.c
@@ -1,6 +1,6 @@
 #include "Colors.h"
 
-const u32 lcd_colors[LCD_COLOR_COUNT] = {
+const u16 lcd_colors[LCD_COLOR_COUNT] = {
                                         WHITE,
                                         BLACK,
                                         RED,
